+++ example/lib/data/stock_repository.dart
@@ -1,5 +1,4 @@
 import 'dart:async';
-import 'dart:ui';
 
 import 'package:equatable/equatable.dart';
 // ignore: implementation_imports