+++ example/lib/pages/simple/huge_data.dart
@@ -1,8 +1,8 @@
 import 'package:example/data/big_data_source.dart';
 import 'package:ff_annotation_route_library/ff_annotation_route_library.dart';
-import 'package:flutter/material.dart';
 @FFArgumentImport()
 import 'package:flex_grid/flex_grid.dart';
+import 'package:flutter/material.dart';
 import 'package:pull_to_refresh_notification/pull_to_refresh_notification.dart';
 
 @FFRoute(