+++ lib/flex_grid.dart
@@ -1,12 +1,8 @@
 library flex_grid;
 
+export 'package:extended_tabs/extended_tabs.dart';
 export 'src/collection/list.dart';
 export 'src/collection/row.dart';
-
-export 'src/controller/drag_hold_controller.dart';
-export 'src/controller/scroll_controller.dart';
-export 'src/controller/scroll_physics.dart';
 export 'src/flex_grid.dart';
-export 'src/horizontal_sync_scroll_minxin.dart';
 export 'src/style/style.dart';
 export 'src/typedef.dart';